Replaced String filePath with File file.
Original design would have forced playSong to directly access the database in order to play the mp3, violating clean architecture.

A better alternative might be a Media object if the only purpose the File serves is to be played, but 1) I don't know if anyone else needs to access the file and 2) I'm not certain which audio-playing library @Mistryr222 plans on using. From what I can tell, the common method (which uses Media) requires javafx to be installed.

I'm committing this as a File for now just to ensure nobody uses the original string fileLocation.